apply fixes for: https://issues.apache.org/jira/browse/AMQCPP-385
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@1213416 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/transport/correlator/ResponseCorrelator.cpp
@@ -103,7 +103,7 @@ Pointer<Response> ResponseCorrelator::request( const Pointer<Command>& command )
         Pointer<FutureResponse> futureResponse( new FutureResponse() );
 
         synchronized( &mapMutex ){
-            requestMap.insert( make_pair( command->getCommandId(), futureResponse ) );
+            requestMap.insert( make_pair( (unsigned int)command->getCommandId(), futureResponse ) );
         }
 
         // The finalizer will cleanup the map even if an exception is thrown.
@@ -147,7 +147,7 @@ Pointer<Response> ResponseCorrelator::request( const Pointer<Command>& command, 
         Pointer<FutureResponse> futureResponse( new FutureResponse() );
 
         synchronized( &mapMutex ){
-            requestMap.insert( make_pair( command->getCommandId(), futureResponse ) );
+            requestMap.insert( make_pair( (unsigned int)command->getCommandId(), futureResponse ) );
         }
 
         // The finalizer will cleanup the map even if an exception is thrown.
